♻️ refactor: move to test_block_access_lists
