--- tests/amsterdam/eip7928_block_level_access_lists/test_block_access_lists_self_destruct.py
@@ -1,134 +1,0 @@
-"""Tests for effects of `SELFDESTRUCT` on EIP-7928 ."""
-
-import pytest
-
-from ethereum_test_tools import (
-    Account,
-    Alloc,
-    BalCodeChange,
-    Block,
-    BlockchainTestFiller,
-    Initcode,
-    Transaction,
-    compute_create_address,
-)
-from ethereum_test_types.block_access_list import (
-    BalAccountExpectation,
-    BalBalanceChange,
-    BalNonceChange,
-    BlockAccessListExpectation,
-)
-from ethereum_test_vm import Opcodes as Op
-
-from .spec import ref_spec_7928
-
-REFERENCE_SPEC_GIT_PATH = ref_spec_7928.git_path
-REFERENCE_SPEC_VERSION = ref_spec_7928.version
-
-
-pytestmark = pytest.mark.valid_from("Amsterdam")
-
-
-@pytest.mark.parametrize(
-    "self_destruct_in_same_tx",
-    [True, False],
-    ids=["self_destruct_in_same_tx", "self_destruct_in_a_new_tx"],
-)
-def test_bal_self_destruct(
-    pre: Alloc,
-    blockchain_test: BlockchainTestFiller,
-    self_destruct_in_same_tx: bool,
-):
-    """Ensure BAL captures balance changes caused by `SELFDESTRUCT`."""
-    alice = pre.fund_eoa()
-    bob = pre.fund_eoa(amount=0)
-
-    selfdestruct_code = Op.SELFDESTRUCT(bob)
-    # A pre existing self-destruct contract
-    kaboom = pre.deploy_contract(code=selfdestruct_code)
-
-    # A template for self-destruct contract
-    self_destruct_init_code = Initcode(deploy_code=selfdestruct_code)
-    template = pre.deploy_contract(code=self_destruct_init_code)
-
-    if self_destruct_in_same_tx:
-        # The goal is to create a self-destructing contract in the same
-        # transaction to trigger deletion of code as per EIP-6780.
-        # The factory contract below creates a new self-destructing
-        # contract and calls it in this transaction.
-
-        bytecode_size = len(self_destruct_init_code)
-        factory_bytecode = (
-            # Clone template memory
-            Op.EXTCODECOPY(template, 0, 0, bytecode_size)
-            # Fund 100 wei and deploy the clone
-            + Op.CREATE(100, 0, bytecode_size)
-            # Call the clone, which self-destructs
-            + Op.CALL(50_000, Op.DUP6, 0, 0, 0, 0, 0)
-            + Op.STOP
-        )
-
-        factory = pre.deploy_contract(code=factory_bytecode)
-        kaboom_same_tx = compute_create_address(address=factory, nonce=1)
-
-    tx = Transaction(
-        sender=alice,
-        to=factory if self_destruct_in_same_tx else kaboom,
-        value=100,
-        gas_limit=1_000_000,
-    )
-
-    # Determine which account was destructed
-    self_destructed_account = kaboom_same_tx if self_destruct_in_same_tx else kaboom
-
-    block = Block(
-        txs=[tx],
-        expected_block_access_list=BlockAccessListExpectation(
-            account_expectations={
-                alice: BalAccountExpectation(
-                    nonce_changes=[BalNonceChange(tx_index=1, post_nonce=1)],
-                ),
-                bob: BalAccountExpectation(
-                    balance_changes=[BalBalanceChange(tx_index=1, post_balance=100)]
-                ),
-                self_destructed_account: BalAccountExpectation(
-                    # When an account is deleted its post nonce is set to 0
-                    # as per EIP-7928.
-                    nonce_changes=[BalNonceChange(tx_index=1, post_nonce=0)]
-                    if self_destruct_in_same_tx
-                    else [],
-                    balance_changes=[BalBalanceChange(tx_index=1, post_balance=0)],
-                    # Expect code to be cleared if self-destructed in same transaction.
-                    code_changes=[BalCodeChange(tx_index=1, new_code="")]
-                    if self_destruct_in_same_tx
-                    else [],
-                ),
-            }
-        ),
-    )
-
-    post = {
-        alice: Account(nonce=1),
-        bob: Account(balance=100),
-        kaboom: Account(balance=0, code=selfdestruct_code),
-    }
-
-    # If the account was self-destructed in the same transaction,
-    # we expect the account to non-existent and its balance to be 0.
-    if self_destruct_in_same_tx:
-        post.update(
-            {
-                factory: Account(
-                    nonce=2,  # incremented after CREATE
-                    balance=0,  # spent on CREATE
-                    code=factory_bytecode,
-                ),
-                kaboom_same_tx: Account.NONEXISTENT,  # type: ignore
-            }
-        )
-
-    blockchain_test(
-        pre=pre,
-        blocks=[block],
-        post=post,
-    )
